Delete d'un seul favoris d'un user

+++ src/main/java/com/rossbach/controller/FavorisController.java
@@ -3,6 +3,7 @@ package com.rossbach.controller;
 import com.rossbach.entities.Favoris;
 import com.rossbach.service.FavorisService;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -36,8 +37,32 @@ public class FavorisController {
         return this.favorisService.showFavorisByID(id);
     }
 
-    @DeleteMapping(path = "/user/{userId}")
-    public void deleteFavorisByUserId(@PathVariable int userId){
+    @DeleteMapping(path = "/user/{userId}/all")
+    public ResponseEntity<String> deleteFavorisByUserId(@PathVariable int userId){
+        boolean userIdExistInVue = this.favorisService.checkIfUserIdExistsInFavoris(userId);
+        if(!userIdExistInVue){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body("Aucun favoris n'est associée à l'utilisateur avec l'ID " + userId);
+        }
         this.favorisService.deleteFavorisByUserId(userId);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Les favoris associées à l'utilisateur avec l'ID \" + userId + \" ont été supprimées avec succès\"");
     }
+
+    @DeleteMapping(path = "{favorisId}/user/{userId}")
+    public ResponseEntity<String> deleteSingleVueByUserId(@PathVariable("userId") int userId, @PathVariable("favorisId") int favorisId) {
+        // Vérifier si la vue appartient à l'utilisateur
+        boolean favorisBelongsToUser = this.favorisService.checkIfFavorisBelongsToUser(userId, favorisId);
+        if (!favorisBelongsToUser) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .body("La vue spécifiée n'appartient pas à l'utilisateur avec l'ID " + userId);
+        }
+
+        // Supprimer la vue spécifique
+        this.favorisService.deleteSingleFavorisByUserId(userId, favorisId);
+
+        return ResponseEntity.status(HttpStatus.OK)
+                .body("La vue avec l'ID " + favorisId + " de l'utilisateur avec l'ID " + userId + " a été supprimée avec succès");
+    }
+
 }
